Replace the Simple Calendar plugin with a native calendar handler
- Remove the Simple Calendar plugin and its registration
- Add a handler that refreshes merchants and hides closed tokens on world time changes
- Initialize the handler on ready whenever a world calendar is present
